requirements.txt
streamlit
tensorflow==2.14.0
numpy
pillow
